cleanest file thus far
This is the version that will stay the cleanest and be updated all the time.  So, every other version is more about saving things that maybe needed, but this one includes the cleanest work from A to Z. So it starts by uploading and cleaning the data, to catching high correlations, to running a simple scatterplot for demonstration.
Next is about to divide data into training and testing and getting ready for classifiers.