Update
Various classifiers are tested in sampling sizes that vary from sampling size = 100 to sampling size = 300. The idea is to identify the classifier that works best with this dataset in predicting the kind of breast cancer for this population. It seems that Quadratic Discriminant Analysis provides the highest score in predicting the type of cancer. The F-1 score is highest in this case.  The F-1 score is a measure of accuracy or the ratio of the data that was accurately predicted. A score of 1 indicates the perfect prediction and a score of 0 shows the worst prediction.